+++ src/components/revision_files.rs
@@ -11,14 +11,17 @@ use crate::{
 	AsyncAppNotification, AsyncNotification,
 };
 use anyhow::Result;
-use asyncgit::sync::{
-	self, get_commit_info, CommitId, CommitInfo, RepoPathRef,
-	TreeFile,
+use asyncgit::{
+	asyncjob::AsyncSingleJob,
+	sync::{
+		get_commit_info, CommitId, CommitInfo, RepoPathRef, TreeFile,
+	},
+	AsyncGitNotification, AsyncTreeFilesJob,
 };
 use crossbeam_channel::Sender;
 use crossterm::event::Event;
 use filetreelist::{FileTree, FileTreeItem};
-use std::fmt::Write;
+use std::{borrow::Cow, fmt::Write};
 use std::{
 	collections::BTreeSet,
 	convert::From,
@@ -44,7 +47,8 @@ pub struct RevisionFilesComponent {
 	queue: Queue,
 	theme: SharedTheme,
 	//TODO: store TreeFiles in `tree`
-	files: Vec<TreeFile>,
+	files: Option<Vec<TreeFile>>,
+	async_treefiles: AsyncSingleJob<AsyncTreeFilesJob>,
 	current_file: SyntaxTextComponent,
 	tree: FileTree,
 	scroll: VerticalScroll,
@@ -60,6 +64,7 @@ impl RevisionFilesComponent {
 		repo: RepoPathRef,
 		queue: &Queue,
 		sender: &Sender<AsyncAppNotification>,
+		sender_git: Sender<AsyncGitNotification>,
 		theme: SharedTheme,
 		key_config: SharedKeyConfig,
 	) -> Self {
@@ -73,8 +78,9 @@ impl RevisionFilesComponent {
 				key_config.clone(),
 				theme.clone(),
 			),
+			async_treefiles: AsyncSingleJob::new(sender_git),
 			theme,
-			files: Vec::new(),
+			files: None,
 			revision: None,
 			focus: Focus::Tree,
 			key_config,
@@ -89,13 +95,15 @@ impl RevisionFilesComponent {
 
 		let same_id =
 			self.revision.as_ref().map_or(false, |c| c.id == commit);
+
 		if !same_id {
-			self.files =
-				sync::tree_files(&self.repo.borrow(), commit)?;
-			let filenames: Vec<&Path> =
-				self.files.iter().map(|f| f.path.as_path()).collect();
-			self.tree = FileTree::new(&filenames, &BTreeSet::new())?;
-			self.tree.collapse_but_root();
+			self.files = None;
+
+			self.async_treefiles.spawn(AsyncTreeFilesJob::new(
+				self.repo.borrow().clone(),
+				commit,
+			));
+
 			self.revision =
 				Some(get_commit_info(&self.repo.borrow(), &commit)?);
 		}
@@ -114,13 +122,35 @@ impl RevisionFilesComponent {
 	}
 
 	///
-	pub fn update(&mut self, ev: AsyncNotification) {
+	pub fn update(&mut self, ev: AsyncNotification) -> Result<()> {
 		self.current_file.update(ev);
+
+		if matches!(
+			ev,
+			AsyncNotification::Git(AsyncGitNotification::TreeFiles)
+		) {
+			if let Some(last) = self.async_treefiles.take_last() {
+				if let Some(Ok(last)) = last.result() {
+					let filenames: Vec<&Path> = last
+						.iter()
+						.map(|f| f.path.as_path())
+						.collect();
+					self.tree =
+						FileTree::new(&filenames, &BTreeSet::new())?;
+					self.tree.collapse_but_root();
+
+					self.files = Some(last);
+				}
+			}
+		}
+
+		Ok(())
 	}
 
 	///
 	pub fn any_work_pending(&self) -> bool {
 		self.current_file.any_work_pending()
+			|| self.async_treefiles.is_pending()
 	}
 
 	fn tree_item_to_span<'a>(
@@ -190,8 +220,9 @@ impl RevisionFilesComponent {
 	}
 
 	fn open_finder(&self) {
-		self.queue
-			.push(InternalEvent::OpenFileFinder(self.files.clone()));
+		self.queue.push(InternalEvent::OpenFileFinder(
+			self.files.clone().unwrap_or_default(),
+		));
 	}
 
 	pub fn find_file(&mut self, file: &Option<PathBuf>) {
@@ -221,18 +252,20 @@ impl RevisionFilesComponent {
 	fn selection_changed(&mut self) {
 		//TODO: retrieve TreeFile from tree datastructure
 		if let Some(file) = self.selected_file_path_with_prefix() {
-			let path = Path::new(&file);
-			if let Some(item) =
-				self.files.iter().find(|f| f.path == path)
-			{
-				if let Ok(path) = path.strip_prefix("./") {
-					return self.current_file.load_file(
-						path.to_string_lossy().to_string(),
-						item,
-					);
+			if let Some(files) = &self.files {
+				let path = Path::new(&file);
+				if let Some(item) =
+					files.iter().find(|f| f.path == path)
+				{
+					if let Ok(path) = path.strip_prefix("./") {
+						return self.current_file.load_file(
+							path.to_string_lossy().to_string(),
+							item,
+						);
+					}
 				}
+				self.current_file.clear();
 			}
-			self.current_file.clear();
 		}
 	}
 
@@ -268,18 +301,30 @@ impl RevisionFilesComponent {
 		let is_tree_focused = matches!(self.focus, Focus::Tree);
 
 		let title = self.title_within(tree_width);
-		ui::draw_list_block(
-			f,
-			area,
-			Block::default()
-				.title(Span::styled(
-					title,
-					self.theme.title(is_tree_focused),
-				))
-				.borders(Borders::ALL)
-				.border_style(self.theme.block(is_tree_focused)),
-			items,
-		);
+		let block = Block::default()
+			.title(Span::styled(
+				title,
+				self.theme.title(is_tree_focused),
+			))
+			.borders(Borders::ALL)
+			.border_style(self.theme.block(is_tree_focused));
+
+		if self.files.is_some() {
+			ui::draw_list_block(f, area, block, items);
+		} else {
+			ui::draw_list_block(
+				f,
+				area,
+				block,
+				vec![Span::styled(
+					Cow::from(strings::loading_text(
+						&self.key_config,
+					)),
+					self.theme.text(false, false),
+				)]
+				.into_iter(),
+			);
+		}
 
 		if is_tree_focused {
 			self.scroll.draw(f, area, &self.theme);